chore: move test helper functions to conftest.py
